refactor: moved fixtures to conftest and common functionality to lib
